Bug persist, but gathered more information
When in edit mode, the row's Select's onChange method correctly updates
the productId. But, the product's name changes visually (in the dropdown),
but the productName state remains the same. Then, when the updateDetail
method is called, a new productId is passed so, of course, the old
product will not be removed (at least with the current implementation).

Some thoughts:
    - Need to make the state of a row detail more consistent, maybe
      gathering all the state into a localDetail state.

    - Either remove the ability to change the product (what product
      is being sold), or add the necessary logic to block addition of a
      repeated product.

    - Make it so the index does not to have to be passed in the way it
      currently is. Better generate it just before each row is created.