Basic wiring for Auth functionality using RTK
The building block is set and tested, but only for the signin (get the
token) functionality, the other parts of the refactoring are still to be
done. Be aware to, that I might want mix and match RTK Query with Slices
and Async Thunks, because these last two would allow me to use code
generated by Swagger codegen.