Fixing 'merge' error
React/Redux/Browser was throwing an error that mentioned a .merge
property being called in an undefined object and that it had something
to do with an api's name. Well, it was caused because I forgot to change
the customersApi reducerPath to 'customers', I had left it at
'managers'. And so, the duplicated API names was what raised the error.