Simplified SideBar's mechanics
the drawerOpen piece of state has been removed as it seemed redundant
with the component prop isOpen. Now, isOpen is the only variable that
controls the state of the drawer. This works, because the sidebar is
re-rendered everytime Base component's sideBarOpen piece of state
changes, when re-rendered the isOpen SideBar prop, will be passed with
another value.