Added custom action resetApp
To hold custom actions, not automaticaly created while creating items in
the reducers section of a Slice, I created the actions.js file at the
root of the store directory.

This actions will allow the application to wipe out the sale list
details/items when the user logsout. So, the actiosn app/resetApp, must
be dispatched at the time of login out of the application.

This actions is yet to be used/dispatched in the application.